[XJC] Create XML Generation project
This is a rudimentary version of the project but it already demonstrates the
core problem: Running `mvn clean generate-source` works on Java 8 but not on
Java 9. That is even if the necessary modules and exports are added to Maven
via `.mvn/jvm.config` (added as `jvm9.config` to keep the project working in
Java 8). The result is "Prefix '' is already bound to ''".