Add empty jvm.configs
If a module folder does not contain an .mvn folder but the parent
folder does, Maven uses the parent .mvn to configure itself. This has
the disadvantage of allowing interference between running the parent
project withh one and the module witn a different Java version.

Best way around that is to create empty .mvn folders but Git doesn't
handle those, so I put an empty .jvm.config in them.